requirements.txt
langchain
flask==3.1.1
flask-cors
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone
langchain-openai
langchain-groq
langchain-community
-e .